fix : logic & add validation on all API

+++ app/controllers/message.controller.js
@@ -1,5 +1,6 @@
 const db = require("../models")
 const moment = require('moment')
+const mongoose = require('mongoose');
 
 const Message = db.message
 
@@ -95,15 +96,25 @@ exports.findAll = async (req, res) => {
 exports.show = async (req, res) => {
     try {
         const { id } = req.params
+
+        const validObjectId = mongoose.Types.ObjectId.isValid(id);
+
+        if (!validObjectId) {
+            return res.status(400).json({
+                message: `Invalid ObjectId: ${id}`
+            });
+        }
+
         const message = await Message.findById(id);
         if (!message) {
             return res.status(400).json({
                 message: `Cannot find any product with ID ${id}`
             })
         }
-        res.status(200).json(message)
+        console.log(message)
+        return res.status(200).json(message)
     } catch (error) {
-       return res.status(500).json({
+        return res.status(500).json({
             message: error.message
         })
     }
@@ -113,9 +124,48 @@ exports.show = async (req, res) => {
 exports.update = async (req, res) => {
     try {
         const { id } = req.params
-        const { type, text, attachment } = req.body;
 
-        
+        const validObjectId = mongoose.Types.ObjectId.isValid(id);
+
+        if (!validObjectId) {
+            return res.status(400).json({
+                message: `Invalid ObjectId: ${id}`
+            });
+        }
+
+        const { type, text, attachment, status } = req.body;
+
+
+        if (type) {
+            if (type !== "text" && type !== "image") {
+                return res.status(400).json({
+                    message: "Type should be filled with 'text' or 'image'"
+                });
+            }
+
+            // Check if the type is "text" and if text is provided
+            if (type === "text" && !text) {
+                return res.status(400).json({
+                    message: 'For type "text", text is required.'
+                });
+            }
+
+            // Check if the type is "image" and if attachmentUrl is provided
+            if (type === "image" && !attachment) {
+                return res.status(400).json({
+                    message: 'For type "image", image is required.'
+                });
+            }
+
+        }
+
+        if (status) {
+            if (status !== "pending" && status !== "delivered" && status !== "sent") {
+                return res.status(400).json({
+                    message: "Status should be filled with 'pending' or 'delivered' or 'sent'"
+                });
+            }
+        }
 
         const message = await Message.findByIdAndUpdate(id, req.body);
 
@@ -126,14 +176,14 @@ exports.update = async (req, res) => {
         }
 
         const updatedMessage = await Message.findById(id)
-        res.status(200).json(
-            {
-                data: updatedMessage
-            }
+        return res.status(200).json({
+            message: `Update message with ID ${id} successful`,
+            data: updatedMessage
+        }
         )
 
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: error.message
         })
     }
@@ -143,17 +193,26 @@ exports.update = async (req, res) => {
 exports.delete = async (req, res) => {
     try {
         const { id } = req.params
+
+        const validObjectId = mongoose.Types.ObjectId.isValid(id);
+
+        if (!validObjectId) {
+            return res.status(400).json({
+                message: `Invalid ObjectId: ${id}`
+            });
+        }
+
         const message = await Message.findByIdAndDelete(id);
         if (!message) {
             return res.status(404).json({
                 message: `Cannot find any message with ID ${id}`
             })
         }
-        res.status(200).json({
+        return res.status(200).json({
             message: `Delete message with ID ${id} successful`
         })
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: error.message
         })
     }
